requirements.txt
annotated-types==0.7.0
anyio==4.10.0
fastapi[standard]
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
psycopg2-binary==2.9.9
uvicorn[standard]==0.32.0
pytest==8.3.3
pytest-asyncio==0.24.0
